extract primitives/load function
when trying to upgrade Next.js a test failed that showed that Uint8Array usage is problematic
with the crypto module. This is because the crypto polyfill is checking `instanceof` of the
input values, and sadly, the Uint8Arrays constructed frmo the EdgeVM context do not satisfy
that condition.

In order to change it, I decided to use the requireWithFakeGlobalScope yet again in the VM
module, but instead of re-implementing it, I extracted the `load` function, which is just
like the `primitives` entrypoint, only lazy--and receives the scopedContext, which will
act as a global context for the requires

The types that are being used in the scopedContext are the same ones we patch for the
`instanceof` checks ourselves, which makes sense!

This fixes the test so we should be safer now!
