+++ packages/primitives/load/package.json
@@ -1,0 +1,4 @@
+{
+  "main": "../dist/load.js",
+  "types": "../types/load.d.ts"
+}
